+++ kernel_tuner/strategies/basinhopping.py
@@ -5,6 +5,8 @@ import scipy.optimize
 
 from kernel_tuner.strategies.minimize import _cost_func, get_bounds_x0_eps, setup_method_arguments, setup_method_options
 
+supported_methods = ["Nelder-Mead", "Powell", "CG", "BFGS", "L-BFGS-B", "TNC", "COBYLA", "SLSQP"]
+
 def tune(runner, kernel_options, device_options, tuning_options):
     """ Find the best performing kernel configuration in the parameter space
 
@@ -30,9 +32,9 @@ def tune(runner, kernel_options, device_options, tuning_options):
     """
 
     results = []
-    cache = {}
 
-    method = tuning_options.method
+    method = tuning_options.strategy_options.get("method", "L-BFGS-B")
+    T = tuning_options.strategy_options.get("T", 1.0)
 
     #scale variables in x to make 'eps' relevant for multiple variables
     tuning_options["scaling"] = True
@@ -43,7 +45,7 @@ def tune(runner, kernel_options, device_options, tuning_options):
     options = setup_method_options(method, tuning_options)
     kwargs['options'] = options
 
-    args = (kernel_options, tuning_options, runner, results, cache)
+    args = (kernel_options, tuning_options, runner, results)
 
     minimizer_kwargs = dict(**kwargs)
     minimizer_kwargs["method"] = method