+++ kernel_tuner/util.py
@@ -1,6 +1,7 @@
 """ Module for kernel tuner utility functions """
 from __future__ import print_function
 
+import json
 from collections import OrderedDict
 import os
 import errno
@@ -8,6 +9,8 @@ import tempfile
 import logging
 import warnings
 import re
+
+import pytest
 import numpy
 
 default_block_size_names = ["block_size_x", "block_size_y", "block_size_z"]
@@ -377,3 +380,99 @@ def normalize_verify_function(v):
     if has_kw_argument(v, 'atol'):
         return v
     return lambda answer, result_host, atol: v(answer, result_host)
+
+
+def process_cache(cache, kernel_options, tuning_options, runner):
+    """cache file for storing tuned configurations
+
+    the cache file is stored using JSON and uses the following format:
+
+    { device_name: "name of device"
+      kernel_name: "name of kernel"
+      tune_params_keys: list
+      tune_params:
+      cache: {
+      "x1,x2,..xN": {"block_size_x": x1, ..., time=0.234342},
+      "y1,y2,..yN": {"block_size_x": y1, ..., time=0.134233},
+      }
+    }
+
+    The last two closing brackets "}\n}" are not required, and everything
+    should work as expected if these are missing. This is to allow to continue
+    from an earlier (abruptly ended) tuning session.
+
+    """
+    #caching only works correctly if tunable_parameters are stored in a OrderedDict
+    if not isinstance(tuning_options.tune_params, OrderedDict):
+        raise ValueError("Caching only works correctly when tunable parameters are stored in a OrderedDict")
+
+    #if file does not exist, create new cache
+    if not os.path.isfile(cache):
+        c = OrderedDict()
+        c["device_name"] = runner.dev.name
+        c["kernel_name"] = kernel_options.kernel_name
+        c["tune_params_keys"] = list(tuning_options.tune_params.keys())
+        c["tune_params"] = tuning_options.tune_params
+        c["cache"] = {}
+
+        contents = json.dumps(c, indent="")[:-3] #except the last "}\n}"
+
+        #write the header to the cachefile
+        with open(cache, "w") as cachefile:
+            cachefile.write(contents)
+
+        tuning_options.cachefile = cache
+        tuning_options.cache = {}
+
+    #if file exists
+    else:
+        with open(cache, "r") as cachefile:
+            filestr = cachefile.read().strip()
+
+        #if file was not properly closed, pretend it was properly closed
+        if not filestr[-3:] == "}\n}":
+            #remove the trailing comma if any, and append closing brackets
+            if filestr[-1] == ",":
+                filestr = filestr[:-1]
+            filestr = filestr + "}\n}" 
+        else:
+            #if it was properly closed, open it for appending new entries
+            with open(cache, "w") as cachefile:
+                cachefile.write(filestr[:-3] + ",")
+
+        cached_data = json.loads(filestr)
+
+        #check if it is safe to continue tuning from this cache
+        if cached_data["device_name"] != runner.dev.name:
+            raise ValueError("Cannot load cache which contains results for different device")
+        if cached_data["kernel_name"] != kernel_options.kernel_name:
+            raise ValueError("Cannot load cache which contains results for different kernel")
+        if cached_data["tune_params_keys"] != list(tuning_options.tune_params.keys()):
+            raise ValueError("Cannot load cache which contains results obtained with different tunable parameters")
+
+        tuning_options.cachefile = cache
+        tuning_options.cache = cached_data["cache"]
+
+
+def close_cache(cache):
+    if not os.path.isfile(cache):
+        raise ValueError("close_cache expects cache file to exist")
+
+    with open(cache, "r") as fh:
+        contents = fh.read()
+
+    #close to file to make sure it can be read by JSON parsers
+    if contents[-1] == ",":
+        with open(cache, "w") as fh:
+            fh.write(contents[:-1] + "}\n}")
+
+
+def store_cache(key, params, tuning_options):
+    logging.debug('store_cache called, cache=%s, cachefile=%s' % (tuning_options.cache, tuning_options.cachefile))
+    if isinstance(tuning_options.cache, dict):
+        if not key in tuning_options.cache:
+            tuning_options.cache[key] = params
+            if tuning_options.cachefile:
+                with open(tuning_options.cachefile, "a") as cachefile:
+                    cachefile.write("\n" + json.dumps({key: params})[1:-1] + ",")
+